chore: more restructuring
