+++ packages/instant-apps-components/src/components/instant-apps-time-filter/TimeFilter/TimeFilterViewModel.ts
@@ -1,4 +1,4 @@
-import { state } from './model';
+import { state } from './TimeFilterModel';
 import { loadModules } from '../../../utils/loadModules';
 
 import { DateValue, ITimeInfoConfigItem, ITimeInfoItem, ITimeItemUnit } from './interfaces/interfaces';
